conditionalize the server scripts in /setup experience so non raspberrypi setups don't fail with a internal server error

+++ app/main/routes_server.py
@@ -15,7 +15,7 @@ from os import path
 
 from . import main
 from .config import base_path
-from .frontend_common import render_template_with_defaults, system_info
+from .frontend_common import platform, render_template_with_defaults, system_info
 
 
 # -------- Routes --------
@@ -35,13 +35,19 @@ def restart_server():
 
 @main.route('/restart_system')
 def restart_system():
+    if platform() != "RaspberryPi":
+        return '', 404
+    
     os.system('shutdown -r now')
     # TODO: redirect to a page with alert of restart
     return redirect('/')
-
+    
 
 @main.route('/shutdown_system')
 def shutdown_system():
+    if platform() != "RaspberryPi":
+        return '', 404
+    
     os.system('shutdown -h now')
     # TODO: redirect to a page with alert of shutdown
     return redirect('/')
@@ -49,11 +55,17 @@ def shutdown_system():
 
 @main.route('/logs')
 def view_logs():
+    if platform() != "RaspberryPi":
+        return '', 404
+    
     return render_template_with_defaults('logs.html')
-
+    
 
 @main.route('/logs/<log_type>.log')
 def download_logs(log_type):
+    if platform() != "RaspberryPi":
+        return '', 404
+
     try:
         filename = ""
         if log_type == 'nginx.access':
@@ -99,6 +111,9 @@ def available_networks():
 @main.route('/setup', methods=['GET', 'POST'])
 def setup():
     if request.method == 'POST':
+        if platform() != "RaspberryPi":
+            return '', 404
+    
         payload = request.get_json()
 
         if 'hostname' in payload:
@@ -209,20 +224,25 @@ def setup():
             current_app.logger.error("ERROR: unsupported payload received %s".format(payload))
             return 'Invalid Setup Payload Received - Setup Failed!', 418
     else:
-        return render_template_with_defaults('setup.html',
-            hostname=hostname(),
-            ap0=accesspoint_credentials(),
-            wireless_credentials=wireless_credentials(),
-            available_networks=available_networks())
+        if platform() == "RaspberryPi":
+            return render_template_with_defaults('setup.html',
+                hostname=hostname(),
+                ap0=accesspoint_credentials(),
+                wireless_credentials=wireless_credentials(),
+                available_networks=available_networks())
+        else:
+            return render_template_with_defaults('setup.html',
+                hostname=hostname(),
+                ap0=None,
+                wireless_credentials=None,
+                available_networks=None)
 
 
 def hostname():
     try:
-        subprocess.check_output("more /etc/hostname", shell=True)
-        return subprocess.check_output("hostname", shell=True).decode("utf-8").strip()
+        return subprocess.check_output("more /etc/hostname || hostname", shell=True).decode("utf-8").strip()
     except:
         current_app.logger.warn("current device doesn't support hostname changes")
-
     return None
